+++ lib/wled.dart
@@ -1,2 +1,3 @@
 export 'src/wled.dart' show Wled;
 export 'src/status.dart' show WledStatus;
+export 'src/color.dart' show Color, Colors;